fixed according to comments
- separate files for controls & validation
- check body when POST & PUT
