+++ routes/api/api.js
@@ -1,0 +1,68 @@
+const contacts = require("../../models/contacts");
+const { HttpError } = require("../../helpers");
+
+const fetchListContacts = async (req, res, next) => {
+  try {
+    const allContacts = await contacts.listContacts();
+    res.json(allContacts);
+  } catch (error) {
+    next(error);
+  }
+};
+
+const fetchContact = async (req, res, next) => {
+  try {
+    const { contactId } = req.params;
+    const oneContact = await contacts.getContactById(contactId);
+    if (!oneContact) {
+      throw HttpError(404, "Not Found");
+    }
+    res.json(oneContact);
+  } catch (error) {
+    next(error);
+  }
+};
+
+const addContact = async (req, res, next) => {
+  try {
+    const addedContact = await contacts.addContact(req.body);
+    res.status(201).json(addedContact);
+  } catch (error) {
+    next(error);
+  }
+};
+
+const deleteContact = async (req, res, next) => {
+  try {
+    const { contactId } = req.params;
+    const deletedContact = await contacts.removeContact(contactId);
+    if (!deletedContact) {
+      throw HttpError(404, "Not Found");
+    }
+    console.log("removed:", deletedContact);
+    res.json({ message: "contact deleted" });
+  } catch (error) {
+    next(error);
+  }
+};
+
+const changeContact = async (req, res, next) => {
+  try {
+    const { contactId } = req.params;
+    const updatedContact = await contacts.updateContact(contactId, req.body);
+    if (!updatedContact) {
+      throw HttpError(404, "Not Found");
+    }
+    res.json(updatedContact);
+  } catch (error) {
+    next(error);
+  }
+};
+
+module.exports = {
+  fetchListContacts,
+  fetchContact,
+  addContact,
+  deleteContact,
+  changeContact,
+};
