+++ routes/api/contacts.js
@@ -1,86 +1,23 @@
 const express = require("express");
-const Joi = require("joi");
-
 const router = express.Router();
-const contactsAddSchema = Joi.object({
-  name: Joi.string().required().messages({
-    "any.required": `missing required name field`,
-  }),
-  email: Joi.string().required().messages({
-    "any.required": `missing required email field`,
-  }),
-  phone: Joi.string().required().messages({
-    "any.required": `missing required phone field`,
-  }),
-});
 
-const contacts = require("../../models/contacts");
-const { HttpError } = require("../../helpers");
+const validateData = require("./validation");
+const {
+  fetchListContacts,
+  fetchContact,
+  addContact,
+  deleteContact,
+  changeContact,
+} = require("./api");
 
-router.get("/", async (req, res, next) => {
-  try {
-    const allContacts = await contacts.listContacts();
-    res.json(allContacts);
-  } catch (error) {
-    next(error);
-  }
-});
+router.get("/", fetchListContacts);
 
-router.get("/:contactId", async (req, res, next) => {
-  try {
-    const { contactId } = req.params;
-    const oneContact = await contacts.getContactById(contactId);
-    if (!oneContact) {
-      throw HttpError(404, "Not Found");
-    }
-    res.json(oneContact);
-  } catch (error) {
-    next(error);
-  }
-});
+router.get("/:contactId", fetchContact);
 
-router.post("/", async (req, res, next) => {
-  try {
-    const { error } = contactsAddSchema.validate(req.body);
-    if (error) {
-      throw HttpError(400, error.message);
-    }
-    const addedContact = await contacts.addContact(req.body);
-    res.status(201).json(addedContact);
-  } catch (error) {
-    next(error);
-  }
-});
+router.post("/", validateData, addContact);
 
-router.delete("/:contactId", async (req, res, next) => {
-  try {
-    const { contactId } = req.params;
-    const deletedContact = await contacts.removeContact(contactId);
-    if (!deletedContact) {
-      throw HttpError(404, "Not Found");
-    }
-    console.log("removed:", deletedContact);
-    res.json({ message: "contact deleted" });
-  } catch (error) {
-    next(error);
-  }
-});
+router.delete("/:contactId", deleteContact);
 
-router.put("/:contactId", async (req, res, next) => {
-  try {
-    const { error } = contactsAddSchema.validate(req.body);
-    if (error) {
-      throw HttpError(400, error.message);
-    }
-    const { contactId } = req.params;
-    const updatedContact = await contacts.updateContact(contactId, req.body);
-    if (!updatedContact) {
-      throw HttpError(404, "Not Found");
-    }
-    res.json(updatedContact);
-  } catch (error) {
-    next(error);
-  }
-});
+router.put("/:contactId", validateData, changeContact);
 
 module.exports = router;
